delete

+++ src/app/devstudio/customersOrder/ordersList/ordersList.controller.js
@@ -1,25 +1,27 @@
 (function () {
     angular.module("orderCustomerModule",[])
-        .controller("orderCustomerCtrl", orderCustomerCtrl)
+        .controller("orderCustomerCtrl", orderCustomerCtrl);
 
     orderCustomerCtrl.$inject = ["$scope", "serverDataService", "FileSaver", "Blob", "$templateCache", "$compile", "$timeout", "serverActService"];
     function orderCustomerCtrl($scope, serverDataService, FileSaver, Blob, $templateCache, $compile, $timeout, serverActService){
-        var self = this;
-        $scope.itemsByPage = 10;
+        var vm = this;
+        vm.itemsByPage = 10;
 
-        $scope.deleteCustomerOrder = function(order){
+        vm.deleteCustomerOrder = function(order){
+            var checkDelete = confirm("Видалити замовлення");
+            if(!checkDelete) return;
             serverActService.deleteCustomerOrder(order).then(function (response) {
-                  alert("ahueno");
+                getCustomers();
             });
         };
 
-        $scope.exportData = function () {
+        vm.exportData = function () {
             if(!document.getElementById('fullOrderCustomersTable')) {
                 var el = document.createElement("div");
                 el.innerHTML = $templateCache.get("devstudio/customersOrder/orderTable/fullOrderCustomersTable.tmpl.html");
                 el.style = "display: none";
                 var compiled = $compile(el);
-                compiled($scope);
+                compiled(vm);
                 document.body.append(el);
             }
 
@@ -27,22 +29,21 @@
                 var blob = new Blob([document.getElementById('fullOrderCustomersTable').innerHTML], {
                     type: "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet;"
                 });
-                FileSaver.saveAs(blob, "Report.xlsx");
+                FileSaver.saveAs(blob, "Report.xls");
             }, 100);
         };
 
-        $scope.saveData = function (data){
+        vm.saveData = function (data){
             serverDataService.getCustomerItem(data.id).then(
                 function(data){
-                    $scope.order = data;
+                    $scope.order = vm.order = data;
                     if(!document.getElementById('exportOrder')) {
-                        console.log($scope.order);
                         var el = document.createElement("div");
                         el.innerHTML = $templateCache.get("devstudio/customersOrder/orderTable/orderCustomerTable.tmpl.html");
                         el.style = "display: none";
-                        var compiled = $compile(el);
-                        compiled($scope);
-                        document.body.append(el);
+                        // var compiled = $compile(el);
+                        // compiled($scope);
+                        document.getElementById('exportTable').append(el);
                     }
                     $timeout(function () {
                         var  blob = new Blob([document.getElementById('exportOrder').innerHTML], {type: "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet;"});
@@ -53,19 +54,17 @@
             )
         };
 
-        self.getCustomers = function(){
+        function getCustomers(){
             serverDataService.getCustomers().then(
                 function(data){
-                    $scope.customers = [];
-                    $scope.customers = data;
-                    $scope.customerCollection = [].concat($scope.customers);
+                    $scope.customers = vm.customers = [];
+                    $scope.customers = vm.customers = data;
+                    $scope.customerCollection = vm.customerCollection = [].concat(vm.customers);
                 }
             )
         };
 
-        (function(){
-            self.getCustomers();
-        })();
+        getCustomers();
     }
 })();
